requirements.txt
google-api-python-client
gcloud
pycryptodome
google-auth
google-auth-httplib2